do not escape links in sourcecode without [source] header: https://github.com/metanorma/iso-10303-detached-docs/issues/239
