+++ lib/metanorma/standoc/blocks.rb
@@ -1,5 +1,6 @@
 require "uri" if /^2\./.match?(RUBY_VERSION)
 require_relative "./blocks_notes"
+require_relative "./blocks_image"
 
 module Metanorma
   module Standoc
@@ -119,25 +120,6 @@ module Metanorma
                     @reqt_models.requirement_roles[role.to_sym], role)
       end
 
-      def svgmap_attrs(node)
-        attr_code(id_attr(node)
-          .merge(id: node.id, number: node.attr("number"),
-                 unnumbered: node.option?("unnumbered") ? "true" : nil,
-                 subsequence: node.attr("subsequence"))
-        .merge(keep_attrs(node)))
-      end
-
-      def svgmap_example(node)
-        noko do |xml|
-          xml.svgmap **attr_code(svgmap_attrs(node).merge(
-                                   src: node.attr("src"), alt: node.attr("alt"),
-                                 )) do |ex|
-            figure_title(node, ex)
-            ex << node.content
-          end
-        end.join("\n")
-      end
-
       # prevent A's and other subs inappropriate for pseudocode
       def pseudocode_example(node)
         node.blocks.each { |b| b.remove_sub(:replacements) }
@@ -160,34 +142,6 @@ module Metanorma
             wrap_in_para(node, ex)
           end
         end.join("\n")
-      end
-
-      def figure_example(node)
-        noko do |xml|
-          xml.figure **figure_attrs(node) do |ex|
-            node.title.nil? or ex.name { |name| name << node.title }
-            wrap_in_para(node, ex)
-          end
-        end.join("\n")
-      end
-
-      def figure_title(node, out)
-        node.title.nil? and return
-        out.name { |name| name << node.title }
-      end
-
-      def figure_attrs(node)
-        attr_code(id_unnum_attrs(node).merge(keep_attrs(node))
-          .merge(class: node.attr("class")))
-      end
-
-      def image(node)
-        noko do |xml|
-          xml.figure **figure_attrs(node) do |f|
-            figure_title(node, f)
-            f.image **image_attributes(node)
-          end
-        end
       end
 
       def para_attrs(node)
